Move Parser::array_values
